network-scheduler upgrade, modifiying examples, fixing bugs
network-scheduler now sends class B donwlink with variable packet size.
Bugs are fixed in the end-device-lora-mac, as well as extended for variable relaying algorithm.
The class-b-network-example-beacon-skippping can be used to analyze beacon skipping, overhearing and throughput for multiple group senario.
The class-b-network-example-beacon-skipping can also be used for simulating unicast, if number of member end-devices per group is 1.
The class-b-network-example-relay is used for analayzing member scalability and do the test for various relaying mechanisms.